split avaliability metrics into own file
